Add custom date formatting by SWLDateFomatter item in Swell.plist file.
The date format is passed to NSDateFormatter to format an NSDate in a log. With this modification, the default date format has been changed to the same format as NSLog because most of case timezone info on NSDate().description is not necessary. Both QuickFormat and Flexformat can be customized to format the date.